+++ src/js/check_in_generator.js
@@ -1,7 +1,7 @@
 
 import PizZip from 'pizzip';
 import Docxtemplater from 'docxtemplater';
-import constants from './constants';
+import constants from './util/constants';
 
 const EMPTY_TEMPLATE_DATA = {
     apartment: "__________________",
