+++ src/js/connector/content-script-connector.js
