--- utils/build.js
@@ -1,11 +1,0 @@
-var webpack = require("webpack"),
-  config = require("../webpack.config");
-
-delete config.chromeExtensionBoilerplate;
-
-webpack(
-  config,
-  function (err) {
-    if (err) throw err;
-  }
-);
